requirements.txt
# Core libraries
numpy
Pillow
opencv-python
scipy

# Image processing and metrics
scikit-image

# Machine learning and deep learning
tensorflow
scikit-learn

# Progress bars (for testing scripts)
tqdm

# API requests (for Unsplash downloader)
requests

# Logging (if not using Python's built-in logging)
logging

# Optional: for Jupyter notebooks if you're using them for development/testing
jupyter
matplotlib

PyWavelets

phe

reedsolo

pyyaml